requirements.txt
pytest==8.3.4
pytest-asyncio==0.25.3
selene==2.0.0rc4
selenium==4.15.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
python-dotenv==1.0.0
requests==2.31.0
